+++ paper/paper.bib
@@ -11,6 +11,8 @@
   pages = {194--210},
   month = jun,
   publisher={Nature Publishing Group}
+  doi = {https://doi.org/10.1038/s41578-018-0026-7},
+  url = {https://www.nature.com/articles/s41578-018-0026-7}
 }
 
 @article{freysoldt2014first,
@@ -21,7 +23,9 @@
   number={1},
   pages={253},
   year={2014},
-  publisher={APS}
+  publisher = {American Physical Society},
+  doi = {10.1103/RevModPhys.86.253},
+  url = {https://link.aps.org/doi/10.1103/RevModPhys.86.253}
 }
 
 @article{stoneham1981non,
@@ -33,6 +37,8 @@
   pages={1251},
   year={1981},
   publisher={IOP Publishing}
+  doi = {10.1088/0034-4885/44/12/001},
+  url = {https://iopscience.iop.org/article/10.1088/0034-4885/44/12/001}  
 }
 
 @article{alkauskas2014first,
@@ -43,7 +49,9 @@
   number={7},
   pages={075202},
   year={2014},
-  publisher={APS}
+  publisher = {American Physical Society},
+  doi = {10.1103/PhysRevB.90.075202},
+  url = {https://link.aps.org/doi/10.1103/PhysRevB.90.075202}
 }
 
 @article{Kim2019kesterite,
@@ -67,4 +75,6 @@ pages = {2686}
   year = {2019},
   month = {Jul},
   publisher = {American Physical Society},
+  doi = {10.1103/PhysRevB.100.041202},
+  url = {https://link.aps.org/doi/10.1103/PhysRevB.100.041202}
 }
