+++ paper/paper.bib
@@ -36,7 +36,7 @@
   number={12},
   pages={1251},
   year={1981},
-  publisher={IOP Publishing}
+  publisher={IOP Publishing},
   doi = {10.1088/0034-4885/44/12/001},
   url = {https://iopscience.iop.org/article/10.1088/0034-4885/44/12/001}  
 }
@@ -61,7 +61,9 @@ journal = {J. Mater. Chem. A},
 year = {2019},
 volume = {7},
 number = {6},
-pages = {2686}
+pages = {2686},
+doi = {10.1039/C8TA10130B},
+url = {https://pubs.rsc.org/en/content/articlelanding/2019/TA/C8TA10130B}
 }
 
 @article{kim2019anham,
