+++ paper/paper.bib
@@ -10,7 +10,7 @@
   number = {7},
   pages = {194--210},
   month = jun,
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
   doi = {https://doi.org/10.1038/s41578-018-0026-7},
   url = {https://www.nature.com/articles/s41578-018-0026-7}
 }
