### 2020-02-11 #2091 @danielskatz
Thanks - one of the DOIs has a leading "https:/doi.org/" that should be removed
