+++ paper/paper.bib
@@ -11,7 +11,7 @@
   pages = {194--210},
   month = jun,
   publisher={Nature Publishing Group},
-  doi = {https://doi.org/10.1038/s41578-018-0026-7},
+  doi = {10.1038/s41578-018-0026-7},
   url = {https://www.nature.com/articles/s41578-018-0026-7}
 }
 
