### 2020-02-13 #2091 @danielskatz
👋 @godotalgorithm & @taylor-a-barnes - would either or both of you be willing to review this submission to JOSS?  JOSS reviews are of the software and short paper, and are checklist-driven, with the goal of improving the software to the point where it can be accepted - see https://joss.readthedocs.io/en/latest/reviewer_guidelines.html for more info
