### 2020-02-13 #2102 @whedon
[REVIEW]: CarrierCapture.jl: Anharmonic Carrier Capture
**Submitting author:** @frssp (<a href="http://orcid.org/0000-0001-5072-6801">Sunghyun Kim</a>)
**Repository:** <a href="https://github.com/WMD-group/CarrierCapture.jl" target ="_blank">https://github.com/WMD-group/CarrierCapture.jl</a>
**Version:** v0.3
