### 2020-02-13 #2102 @danielskatz
Your job now is to examine the paper and the software, and to check off the items on your list that you can, and to explain what blocks you from checking off other items, either here if they are brief, or in new issues in the source repo if they are more complex.  (If you do open issues in the source repo, please mention this review issue in those issues so that we can see them here as well.)

We expect that the review will be done in a couple of weeks, at most, to happen, or at least it will get to a list of blocking items in that timeframe.
