### 2020-02-25 #2102 @taylor-a-barnes
This is a very nice submission with a well-defined scope and general adherence to best practices.  I have just a couple of comments regarding the documentation, which are provided in the above issues.  Otherwise, I recommend that this paper be accepted.
