### 2020-02-26 ef26a6d
Update README.md
Update the installation instruction.
