### 2020-02-26 cab7978
update installation guide
