+++ docs/src/index.md
@@ -17,7 +17,9 @@ The [Brooglie](https://github.com/RedPointyJackson/Brooglie) package is used to 
 Install the package by:
 
 ```
-Pkg.add(PackageSpec(url="https://github.com/WMD-group/CarrierCapture.jl.git"))
+julia> using Pkg
+
+julia> Pkg.add(PackageSpec(url="https://github.com/WMD-group/CarrierCapture.jl.git"))
 ```
 
 
