+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"lib/brooglie/#Public-Documentation-1","page":"Brooglie","title":"Public Documentation","text":"","category":"section"},{"location":"lib/brooglie/#","page":"Brooglie","title":"Brooglie","text":"Documentation for the submodule Brooglie.jl.","category":"page"},{"location":"lib/brooglie/#","page":"Brooglie","title":"Brooglie","text":"This module is a solver for the time-independent Schödinger equation and a copy of a project https://github.com/RedPointyJackson/Brooglie (MIT License).","category":"page"},{"location":"lib/brooglie/#Index-1","page":"Brooglie","title":"Index","text":"","category":"section"},{"location":"lib/brooglie/#","page":"Brooglie","title":"Brooglie","text":"Pages = [\"brooglie.md\"]","category":"page"},{"location":"lib/brooglie/#Public-Interface-1","page":"Brooglie","title":"Public Interface","text":"","category":"section"},{"location":"lib/brooglie/#","page":"Brooglie","title":"Brooglie","text":"Modules = [CarrierCapture.Brooglie]\nPrivate = false","category":"page"},{"location":"lib/brooglie/#CarrierCapture.Brooglie","page":"Brooglie","title":"CarrierCapture.Brooglie","text":"A copy of https://github.com/RedPointyJackson/Brooglie\n\n\n\n\n\n","category":"module"},{"location":"lib/brooglie/#CarrierCapture.Brooglie.buildH-Tuple{Any}","page":"Brooglie","title":"CarrierCapture.Brooglie.buildH","text":"buildH(V; N=20, a=-1, b=1, m=1)\n\nHamiltonian of a particle of mass m in a box spanning from a to b in all D dimensions with a basis of N elements (number of partitions of the space in each coordinate). The potential V(x, y, z, ...) is a function of D arguments.\n\n\n\n\n\n","category":"method"},{"location":"lib/brooglie/#CarrierCapture.Brooglie.solve-Tuple{Any}","page":"Brooglie","title":"CarrierCapture.Brooglie.solve","text":"solve(V; N=500, a=-1, b=1, m=1, nev=N÷20, maxiter=1000)\n\nSolve the potential V(x,y,z,...) in a grid xᵢ ∈ [a,b], discretized in N steps.\n\nThe particle is asumed to have mass m (by default 1, a electron mass).\n\nThe function will return the nev first energy levels (in Hartree[1]) and its normalized eigenfunctions.\n\n[1] A Hartree is equivalent to 27.21… eV. The global variable H2eV, equal to that value, can be accessed under Brooglie.H2eV for convenience.\n\n\n\n\n\n","category":"method"},{"location":"lib/public/#Public-Documentation-1","page":"Public","title":"Public Documentation","text":"","category":"section"},{"location":"lib/public/#","page":"Public","title":"Public","text":"Documentation for CarrierCapture.jl's public interface.","category":"page"},{"location":"lib/public/#Contents-1","page":"Public","title":"Contents","text":"","category":"section"},{"location":"lib/public/#","page":"Public","title":"Public","text":"Pages = [\"public.md\"]","category":"page"},{"location":"lib/public/#Index-1","page":"Public","title":"Index","text":"","category":"section"},{"location":"lib/public/#","page":"Public","title":"Public","text":"Pages = [\"public.md\"]","category":"page"},{"location":"lib/public/#Public-Interface-1","page":"Public","title":"Public Interface","text":"","category":"section"},{"location":"lib/public/#","page":"Public","title":"Public","text":"CarrierCapture\npotential\nfit_pot!\nsolve_pot!\
[…]
