### 2020-02-26 #2102 @danielskatz
Thanks @taylor-a-barnes! I note you didn't check the Functionality documentation checkbox, and that you opened 2 issues in the source repo that are now closed.  Did those changes resolve the overall issue for you, so that you can check off this final checkbox?  If not, what else do you think is needed?
