### 2020-02-26 #2102 @taylor-a-barnes
The sentence at the end of the "Usage" section of the documentation is currently as follows:

> the detailed description provided in the [Usage section](https://wmd-group.github.io/CarrierCapture.jl/dev/usage.html).

This should be corrected for grammatical issues:

> A more detailed description is provided in the [Usage section](https://wmd-group.github.io/CarrierCapture.jl/dev/usage.html).

More importantly, the link is broken.  Once that is fixed, I can check the final checkbox.
