### 2020-02-27 9f99fea
a broken link and a typo are fixed
