### 2020-03-09 863c961
Fix readme typos
