### 2020-03-09 2d3004a
Fix docs typos
