### 2020-03-09 1484c72
Fix typos in API doc
