### 2020-03-09 WMD-group/CarrierCapture.jl#6 @utf
Command line scripts should have help strings
Help strings are very useful for checking what a command line program does without having to open the docs on the website. I.e., so you can type:

```bash
get_cc_struct.py -h
```

Currently, all the tools provided in `CarrierCapture.jl/script/` do not have help strings. It would be nice to add them (even if they are only very short descriptions) to make it clear what the scripts do.

`argparse` makes this very easy. For example:

```python
import argparse

parser = argparse.ArgumentParser(description='This is my help string.')
```

See https://docs.python.org/3.7/library/argparse.html#example for more information.

Part of openjournals/joss-reviews#2102
