### 2020-03-09 WMD-group/CarrierCapture.jl#7 @utf
Example notebooks are not clear
The example notebooks would be a lot more clear if you added some descriptions of what each section is trying to achieve. For someone with no familiarity of the code, I found it quite difficult to parse what each code block was doing. As the notebooks are the main tutorials for the bulk of the code, I think they should be made as clear as possible.

Each notebook should have some discussion at the top, maybe introducing the theory but definitely describing what the notebook will achieve. I would also add the date and version of CarrierCapture.jl used when the notebook was originally run, as this will make it clear when things go out of date.

You can see some excellent example notebooks in these repositories:
- [Matminer examples](https://github.com/hackingmaterials/matminer_examples): In particular, I think this is a good [example](https://github.com/hackingmaterials/matminer_examples/blob/master/matminer_examples/machine_learning-nb/bulk_modulus.ipynb).
- [Bsym](https://github.com/bjmorgan/bsym): Example [here](https://github.com/bjmorgan/bsym/blob/master/examples/bsym_examples.ipynb).

See openjournals/joss-reviews#2102
