### 2020-03-09 WMD-group/CarrierCapture.jl#8 @utf
Who is Antonello Lobianco?
I'm sure he would be very pleased to know CarrierCapture.jl is licensed under his name...

https://github.com/WMD-group/CarrierCapture.jl/blob/9f99feaa840975437e09540ea5d36e6a6e7c3797/LICENCE#L3

Part of openjournals/joss-reviews#2102
