### 2020-03-10 WMD-group/CarrierCapture.jl#9 @utf
Add details of how to run tests
Can you detail somewhere how to run the unit tests, to ensure everything has been installed correctly?

Part of openjournals/joss-reviews#2102
