### 2020-03-10 #2102 @utf
This is a fantastic package and will be of use to many researchers studying first-principles defects. The paper is well written and very clear. The package itself is well structured and the API is clean. The module level documentation is also excellent. Overall, I strongly recommend this paper be accepted.

I have a number of minor concerns which I've raised as pull requests and issues.
