### 2020-03-10 WMD-group/CarrierCapture.jl#8 @frssp
I fixed the `LICENCE` file, before he notices this.
