### 2020-03-10 b22bd2b
Add help strings in python scripts
