### 2020-03-10 WMD-group/CarrierCapture.jl#6 @frssp
The help strings are added.
