### 2020-03-10 WMD-group/CarrierCapture.jl#7 @frssp
Added more descriptions to the notebooks.
