### 2020-03-10 dbbae4c
update doc and README.md; instruction for the test.
