+++ README.md
@@ -27,6 +27,12 @@ julia> using Pkg
 julia> Pkg.add(PackageSpec(url="https://github.com/WMD-group/CarrierCapture.jl.git"))
 ```
 
+To run the unit tests for the package, use the `Pkg.test` function. 
+
+```julia
+julia> Pkg.test("CarrierCapture")
+```
+
 ## Development
 
 The project is hosted on [Github](https://github.com/WMD-group/carriercapture).
