### 2020-03-10 WMD-group/CarrierCapture.jl#9 @frssp
Instruction for the test added.
