+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"usage/#Usage-1","page":"Usage","title":"Usage","text":"","category":"section"},{"location":"usage/#","page":"Usage","title":"Usage","text":"A typical usage will consist of about three steps; 1. preparation, 2. building potential, and 3. computing capture coefficient. Find more detail for step 2 and step 3 in example notebooks. The command line interface is depreciated and not recommended.","category":"page"},{"location":"usage/#.-Preparation-1","page":"Usage","title":"1. Preparation","text":"","category":"section"},{"location":"usage/#","page":"Usage","title":"Usage","text":"Before CarrierCapture, you need to calculate potential energy surfaces of atomic vibrations (one-dimensional Configuration Coordinate diagram; 1D-CC) and e-ph coupling matrix element (W_if). Prepare a sequence of structures with displacements which interpolate between two defect states. Run single-point energy calculations on these structures, and extract the total energies. Scripts for preprocessing can be found in /script which require a python library pymatgen.","category":"page"},{"location":"usage/#","page":"Usage","title":"Usage","text":"Generate 1D-CC\nCalculate equilibirum geometries and total energies of defective supercells with charge states q(initial) and q±1(final) denoted Conf.(q) and Conf.(q±1), respectively.\nGenerate interpolated and extrapolated structures between Conf.(q) (POSCAR_i) and Conf.(q±1) (POSCAR_f). You may use gen_cc_struct.py:\n$ gen_cc_struct.py -i POSCAR_i -f POSCAR_f -d -1 -0.6 -0.4 -0.2 -0.1 0 0.1 0.2 0.4 0.6 1.0\n$ ls\ndisp_dir_i\ndisp_dir_f\nRun total-energy calculations for each structures. Example of the directory tree (template contains all input files for DFT calculations. Make sure DFT-program write wavefunctions (e.g. LWAVE=.TRUE. in VASP) for the next stage W_if):\n├── 00_q2q±1\n│   ├── 00_q\n│   │   ├── DISP_000\n│   │   ├── DISP_001\n│   │   ├── ...\n│   │   ├── disp_dir -> ../../90_DISPLACEMENT/disp_dir_i\n│   │   └── template\n│   └── 01_q±1\n│       ├── DISP_000\n│       ├── DISP_001\n│       ├── ...\n│       ├── disp_dir -> ../../90_DISPLACEMENT/disp_dir_f\n│       └── template\n├── 90_DISPLACEMENT\n│   ├── disp_dir_f\n│   │   ├── POSCAR_000\n│   │   ├── ...\n│   ├── disp_dir_i\n│   │   ├── POSCAR_000\n│   │   ├── ...\nYou can submit jobs for all calculations using a following script in a high-performace computer with a batch system.\n#!/bin/bash -l\n\nfor NUM in {-14,-13,-12,-11,-10,-09,-08,-07,-06,-05,-04,-03,-02,-01,000,001,002,003,004,005,006,007,008,009,010,011,012,013,014}\ndo\n   #if [ ! -d DISP_$NUM ]\n   #then\n     echo DISP_$NUM\n     cp -r template DISP_$NUM\n     cp disp_dir/POSCAR_$NUM DISP_$NUM/POSCAR\n     cd  DISP_$NUM\n     qsub run.pbs\n     cd ../\n   #fi\ndone\nCalculate 𝛥Q using get_del_Q.py. Generate potential.csv using following script.<a name=\"qe_data\"></a>\n#!/bin/bash -l\n\necho Q, E\nfor NUM 
[…]
