### 2020-03-12 #2102 @frssp
- [x] v0.3 is released
- [x] the release v0.3 is archived in Zenodo.
- [x] The Zenodo archive checked.
- [x] The DOI of the archived version is 10.5281/zenodo.3707592
