### 2020-03-12 bf22d60
minor grammar updates
